+++ R/store.R
@@ -447,7 +447,7 @@ ragnar_store_inspect <- function(store, ...) {
 #'
 #' @note This function requires the `embedding-atlas` Python package (>= 0.20.0)
 #' in your reticulate Python environment, the `duckdb` R package (>= 1.4.0),
-#' and `arrow` to transfer data from the DuckDB store to Python.
+#' and `nanoarrow` (>= 0.8.0) to transfer data from the DuckDB store to Python.
 #'
 #' @examples
 #' \dontrun{
@@ -466,11 +466,14 @@ ragnar_store_atlas <- function(
   port = 3030,
   launch.browser = interactive()
 ) {
+  rlang::check_installed("nanoarrow", version = "0.8.0")
+
   # duckdb R version need to match the Python version at least in major.minor
   duckdb_version <- numeric_version(utils::packageVersion("duckdb"))
 
   reticulate::py_require(c(
     "embedding-atlas>=0.20.0",
+    "nanoarrow",
     sprintf("duckdb==%s.%s.*", duckdb_version$major, duckdb_version$minor),
     "numba>=0.62.0rc2",
     "llvmlite>=0.45.0rc2"
@@ -479,11 +482,10 @@ ragnar_store_atlas <- function(
     "_ragnartools.atlas",
     system.file("python", package = "ragnar")
   )
-  df <- duckdb::duckdb_fetch_arrow(DBI::dbSendQuery(
+  df <- DBI::dbGetQueryArrow(
     store@con,
-    "SELECT * FROM chunks;",
-    arrow = TRUE
-  ))
+    "SELECT * FROM chunks;"
+  )
   join_thread <- atlas$run_embedding_atlas(df, host, as.integer(port))
   if (launch.browser) {
     utils::browseURL(sprintf("http://%s:%d", host, port))